feat(strategy) : 전략을 생성할 때, 생성되는 code를 사용자에게 보기좋은 형태로 정리한다
- open ai를 이용해서 코드를 정리한다

+++ strategy-service/src/main/java/com/pda/strategy_service/service/StrategyCodeGenerator.java
@@ -14,13 +14,13 @@ import java.util.*;
 public class StrategyCodeGenerator {
 
   private Set<String> timeframes;
-  private int memberId;
+  private Long memberId;
   private String symbol;
 
   /**
    * 전략 JSON → Python 코드 생성
    */
-  public String generateCode(Map<String, Object> strategyJson, int memberId, String symbol) {
+  public String generateCode(Map<String, Object> strategyJson, Long memberId, String symbol) {
     this.timeframes = analyzeTimeframes(strategyJson);
     this.memberId = memberId;
     this.symbol = symbol;
